requirements.txt
streamlit
pandas
rdkit-pypi
requests
scikit-learn
matplotlib
shap
